Add native systemd service units for dropbear and upmpdcli.
Both dropbear and upmpdcli use a sysv init script but in order to
make use of systemd features (drop-in snippets) we need native
services. This is also a work-around for the bugged version of
systemd (215) that ships with Jessie (see Debian #760616) and does
not support is-enabled for sysv services.

The dropbear unit was adapted from Arch by changing the paths.

The upmpdcli unit was taken from github at
https://github.com/medoc92/upmpdcli/blob/master/systemd/upmpdcli.service